disallow mmap on windows so we can use updated exp pkg.
